Remove local personalization usage and switch IS branding API
